ai_engine/requirements.txt
paho-mqtt==1.6.1
joblib==1.3.2
scikit-learn==1.4.2
numpy==1.26.4
pandas==2.2.2
requests==2.32.3
